# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cloud Edge Backend API",
    "version": "2.4.0",
    "description": "S1-S6 workflow backend and UI-10 management: execution snapshots, guarded logical deletion, ADMIN/USER accounts, image upload/build/distribution, deployments and managed resources. One execution chain; no throughput SDK, long-horizon DQN, physical purge or full Kestra feature parity."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18085"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/api/namespaces/{namespace}/executions/{id}/definition": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "executionDefinition",
        "description": "READ. Immutable execution snapshot, even for removed Flow. Deleted execution returns 404; never current Flow.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowDefinition"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or namespace/action denied"
          },
          "404": {
            "description": "Active record not found"
          },
          "409": {
            "description": "Referenced/active record, revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        }
      }
    },
...
    "/api/namespaces/{namespace}/executions/{id}/output-files": {
      "parameters": [{"name":"namespace","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],
      "get": {"operationId":"executionOutputDeclarations","description":"READ. Container output file declarations from the immutable execution definition, including nested/lifecycle tasks in USER and EDGE_POLICY scopes. Does not expose the complete Flow source or list/read object storage. Actual publication/state remains on TaskRun. No fallback to a current Flow revision.","responses":{"200":{"description":"Declared outputs by task definition ID","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExecutionOutputSource"}}}}},"401":{"description":"Authentication required"},"403":{"description":"Namespace READ denied"},"404":{"description":"Execution not found"}}}
    },
    "/api/namespaces/{namespace}/executions/overview": {"parameters":[{"name":"namespace","in":"path","required":true,"schema":{"type":"string"}}],"get":{"operationId":"executionOverview","description":"READ. Current states of executions submitted in the last 1..31 UTC calendar days, including today up to database time. Full SQL aggregation, not first-page estimates; recent at most 10. No throughput, event-history or workload mutation.","parameters":[{"name":"days","in":"query","schema":{"type":"integer","minimum":1,"maximum":31,"default":7}}],"responses":{"200":{"description":"Window summary","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecutionOverview"}}}},"401":{"description":"Authentication required"},"403":{"description":"Namespace READ denied"},"422":{"description":"Invalid days"}}}},
...
    "/api/namespaces/{namespace}/executions/{id}/tasks/{taskRunId}/output-json": {
      "parameters": [
        {"name": "namespace", "in": "path", "required": true, "schema": {"type": "string"}},
        {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}},
        {"name": "taskRunId", "in": "path", "required": true, "schema": {"type": "string"}}
      ],
      "get": {
        "operationId": "executionOutputJson",
        "description": "READ. Reads an existing JSON output of a SUCCESS TaskRun and its successful Attempt. Port must be declared by the execution's pinned Flow revision. Exact configured artifact bucket and namespace/execution/task/attempt key only. Maximum 262144 bytes; one JSON object; no arbitrary URL, storage listing, ingestion, or writes.",
        "parameters": [{"name": "port", "in": "query", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": {"description": "Original JSON object", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true}}}},
          "401": {"description": "Authentication required"},
          "403": {"description": "Namespace READ denied"},
          "404": {"description": "Execution, task, published port, successful attempt or stored file not found"},
          "409": {"description": "Task has no successfully committed output"},
          "422": {"description": "Undeclared/non-JSON port, foreign artifact URI, oversized or invalid/non-object JSON"},
          "502": {"description": "Object storage unavailable (no credentials disclosed)"}
        }
      }
    },
...
    "/api/namespaces/{namespace}/edge-access/terminals/{id}/executions/{executionId}": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "executionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "terminalResult",
        "description": "CONNECT: existing execution state/outputs/error for this owned terminal only. No ACK or offline queue.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions or ownership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict, disabled resource or different content for key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid definition or inputs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/namespaces/{namespace}/executions": {
      "get": {
        "operationId": "listExecutions",
        "description": "READ: newest createdAt/id first",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "READ: newest createdAt/id first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Execution"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "post": {
        "operationId": "submitExecution",
        "description": "EXECUTE: idempotent by namespace + authenticated actor + key. 202 returned only after commit; repeats return same receipt.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "Stable key for retries of the same request; prefix schedule: is reserved for Scheduler."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "EXECUTE: idempotent by namespace + authenticated actor + key. 202 returned only after commit; repeats return same receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/api/namespaces/{namespace}/executions/{id}": {
      "get": {
        "operationId": "getExecution",
        "description": "READ: persisted execution state and final typed outputs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "READ: persisted execution state and final typed outputs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "delete": {
        "operationId": "removeExecution",
        "description": "WRITE. Only terminal executions with all TaskRuns terminal. Logically hide history/logs/tasks/output reads; retain raw history/idempotency. No cluster/artifact cleanup.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or namespace/action denied"
          },
          "404": {
            "description": "Active record not found"
          },
          "409": {
            "description": "Referenced/active record, revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api/namespaces/{namespace}/executions/{id}/tasks": {
      "get": {
        "operationId": "listTasks",
        "description": "READ: all materialized TaskRuns in creation order, including dynamic scopes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "READ: all materialized TaskRuns in creation order, including dynamic scopes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRun"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/api/namespaces/{namespace}/executions/{id}/tasks/{taskRunId}/attempts": {
      "get": {
        "operationId": "listAttempts",
        "description": "READ: attempt history; retries create attempts, lease takeovers do not",
        "parameters": [],
        "responses": {
          "200": {
            "description": "READ: task run must belong to this execution; ordered attempts include business retries, not lease takeovers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Attempt"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "taskRunId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/api/namespaces/{namespace}/executions/{id}/logs": {
      "get": {
        "operationId": "listLogs",
        "description": "READ: logs in ascending id, afterId exclusive",
        "parameters": [
          {
            "name": "afterId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "READ: logs in ascending id, afterId exclusive",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required; authentication filter response"
          },
          "403": {
            "description": "Namespace/action not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Definition, input or pagination invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Infrastructure failure; request not accepted. Retry submission with original Idempotency-Key."
          }
        }
      },
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/api/namespaces/{namespace}/executions/{id}/cancel": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "cancelExecution",
        "description": "EXECUTE: idempotent cancellation. QUEUED goes directly to KILLED without cleanup. Admitted execution enters KILLING; Finally finishes before KILLED. Terminal execution is unchanged.",
        "responses": {
          "202": {
            "description": "Cancellation accepted or already terminal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "ExecutionOutputSource": {"type":"object","properties":{"id":{"type":"string"},"ports":{"type":"array","items":{"type":"string"}}},"required":["id","ports"]},
      "ExecutionOverview": {"type":"object","properties":{"from":{"type":"string","format":"date-time"},"to":{"type":"string","format":"date-time"},"days":{"type":"array","items":{"$ref":"#/components/schemas/ExecutionDayCount"}},"recent":{"type":"array","items":{"$ref":"#/components/schemas/ExecutionRecent"}}},"required":["from","to","days","recent"]},
      "ExecutionDayCount": {"type":"object","properties":{"date":{"type":"string","format":"date"},"state":{"type":"string","enum":["CREATED","QUEUED","RUNNING","RETRYING","KILLING","SUCCESS","FAILED","KILLED","SKIPPED"]},"count":{"type":"integer","minimum":0}},"required":["date","state","count"]},
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
...
      "ExecutionRequest": {
        "type": "object",
        "properties": {
          "flowId": {
            "type": "string",
            "pattern": "^[A-Za-z][A-Za-z0-9_.-]{0,99}$"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "inputs": {
            "anyOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "flowId"
        ],
        "additionalProperties": false
      },
      "Accepted": {
        "type": "object",
        "properties": {
          "executionId": {
            "type": "string"
          }
        },
        "required": [
          "executionId"
        ],
        "additionalProperties": false
      },
...
    }
  }
}